% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Package.R
\docType{package}
\name{omz12061-package}
\alias{omz12061}
\alias{omz12061-package}
\title{omz12061: A Package Skeleton for Comparative Effectiveness Studies}
\description{
A skeleton package, to be used as a starting point when implementing comparative effect studies.
}
\keyword{internal}
